frontend(ui): add °C / °F toggle in WeatherCard
